Improvements to the submission script
Remove basename because it will print a wrong path if the script is started from another directory, or it is not on the system PATH variable. Also, print help if a wrong number of arguments was given.
Instead of reading the python file, and putting the header into a variable which are then combined, use the cat tool to combine and a heredoc to define the header.